+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Warning log for Reference values in tokens, if used
+
 ## [1.2.0] - 2023-07-26
 ### Added
 - `@fylgja/tailwindcss-plugin-cssprops` as sub package for easier importing,